refactor(docsite): extract shared element resolver, wire slotElements through playground UI
Extracts resolveElementDescriptor and getXDSComponent into a shared
resolveElements.ts module, used by both InteractivePreview (for
playground defaults) and PlaygroundPropsTable (for slot controls).

The slot element flow now works end-to-end:
1. PropDoc declares slotElements: [{__element: 'XDSIcon', props: {...}}]
2. parsePropType sees slotElements → returns kind: 'element' with options
3. ElementControl renders toggle (1 option) or selector (multiple)
4. resolveSlotElement creates the element via resolveElementDescriptor
5. resolveElementDescriptor handles nested children recursively

resolveElementDescriptor looks up components from @xds/core barrel
with both 'Name' and 'XDSName' keys, and falls back to the raw
tag name for non-XDS elements.
